Allow tunnel link to render statically
